Move "shift_the_rest_of_the_line" to a better location.
The function logic is about indentation/placement of chunks, it does not
really belong into "chunk.{h,cpp}".
